Customer Profile And Business info updates
In this commit, the main updates are in the customer profile page and the business information. Now each customer will have their own description, phone number, images, all. So does the business will have the GST number, the PAN number, the business name, description, notes, images, and all. Whatever they are mentioning in OkCredit is already attached in this update.